feat: 프로필 페이지 레이아웃 업데이트
